pyproject.toml
[project]
name = "multiagentplatform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=0.6.3",
    "fastapi>=0.115.12",
    "fastmcp>=2.2.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "ipython>=9.1.0",
    "langchain>=0.3.24",
    "langchain-community>=0.3.22",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.30",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langgraph-cli>=0.2.3",
    "langgraph-supervisor>=0.0.17",
    "langgraph-swarm>=0.0.10",
    "langserve>=0.3.1",
    "langsmith>=0.3.31",
    "psycopg>=3.2.7",
    "psycopg-binary>=3.2.7",
    "psycopg-pool>=3.2.6",
    "pydantic-settings>=2.9.1",
    "pytz>=2025.2",
    "uvicorn>=0.34.0",
    "watchdog>=6.0.0",
    "watchfiles>=1.0.5",
]
